- boolean.cpp: keep HEAD fixes (IsPlanar guard, closed-curve interval check,
  second UV dedup, 1e-10 threshold, orig_face_area coextension guard); take
  booleans shoelace_accumulate (PolylineCurve expansion) and interior-iso
  edge creation; remove CLOSED_CURVE_SAMPLE_COUNT (superseded)
- intersect.cpp: keep HEAD debug-guarded logging
- CMakeLists.txt: use sr_util.cpp (booleans rename)
- .gitignore: merge comprehensive booleans version with brlcad_build/ from HEAD
- nurbs_csg.g: take booleans binary version

+++ brlcad/src/conv/comgeom/read.c
@@ -1,7 +1,7 @@
 /*                          R E A D . C
  * BRL-CAD
  *
- * Copyright (c) 1989-2025 United States Government as represented by
+ * Copyright (c) 1989-2026 United States Government as represented by
  * the U.S. Army Research Laboratory.
  *
  * This program is free software; you can redistribute it and/or
@@ -49,7 +49,7 @@ get_line(char *cp, int buflen, char *title)
     c = bu_vls_gets(&str, infp);
     if (c > buflen)
 	printf("get_line(x%lx, %d) input record overflows buffer for %s\n",
-		(unsigned long)cp, buflen, title);
+		(unsigned long)(uintptr_t)cp, buflen, title);
     int ret = bu_strlcpy(cp, bu_vls_addr(&str), buflen);
     bu_vls_free(&str);
     return ret;
